+++ internal/rpc/settings_commands.go
@@ -53,7 +53,6 @@ func (s *Server) handleSettingsCommand(ctx context.Context, req Request) error {
 			SubagentsMaxConcurrent: params.SubagentsMaxConcurrent,
 			SkillsEnabled:          params.SkillsEnabled,
 			UpdateCheckOnStartup:   params.UpdateCheckOnStartup,
-			AutoUpdate:             params.AutoUpdate,
 		})
 		if err != nil {
 			return err
@@ -73,7 +72,7 @@ func rpcSettings(settings app.SettingsSnapshot) protocol.RPCSettings {
 		SubagentsMaxAgents: settings.SubagentAgentLimit, SkillsEnabled: settings.SkillsEnabled,
 		SubagentsRestartRequired: settings.SubagentsRestartRequired,
 		SkillsRestartRequired:    settings.SkillsRestartRequired, RestartRequired: settings.RestartRequired,
-		UpdateCheckOnStartup: settings.UpdateCheckOnStartup, AutoUpdate: settings.AutoUpdate,
+		UpdateCheckOnStartup: settings.UpdateCheckOnStartup,
 	}
 }